server/CPU/requirements.txt
torch==2.8.0
torchvision==0.23.0
aiohttp==3.12.15
Pillow==11.3.0
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git